requirements.txt
streamlit
torch
torchvision
numpy
pillow
matplotlib